BE-6: made the keyboard shortcut it's own component
